+++ src/router/routes.js
@@ -14,7 +14,6 @@ import pgMenuDetail from 'pages/MenuDetail'
 import pgError404 from 'pages/Error404'
 import pgAdminCustomers from 'pages/admin/Customers'
 import pgTopup from 'pages/Topup'
-import pgStores from 'pages/Stores'
 
 const routes = [
   {
@@ -32,10 +31,6 @@ const routes = [
       {
         path: 'menus/:catId',
         component: pgMenus,
-      },
-      {
-        path: 'stores',
-        component: pgStores,
       },
       {
         path: 'customer',